v1.4.1 修复 OIR ZBC 最大不能超过255的问题 修复OIR内定义了未使用的LUT会导致文件指针不能继续前进的问题

+++ Image5DLib/Oir索引.cpp
@@ -17,7 +17,7 @@ void Oir索引::计算依赖字段(uint32_t 块总数)noexcept
 	if (SizeTZ < SizeZ)
 		SizeZ = SizeTZ;
 	SizeT = SizeTZ / SizeZ;
-	SizeZBC = SizeZ * SizeBC;
+	SizeZBC = (uint16_t)SizeZ * SizeBC;
 	SizeTZBC = SizeT * SizeZBC;
 	SizeYX = (uint32_t)SizeY * SizeX;
 	SizeCYX = SizeC * SizeYX;
